+++ amethyst_assets/src/storage.rs
@@ -60,7 +60,7 @@ where
 {
     /// Asset is not fully loaded yet, need to wait longer
     Loading(A::Data),
-    /// Asset have finished loading, can now be inserted into storage and tracker notified
+    /// Asset has finished loading, can now be inserted into storage and tracker notified
     Loaded(A),
 }
 
@@ -136,7 +136,7 @@ impl<A: Asset> AssetStorage<A> {
         }
     }
 
-    /// Get an asset and it's version from a given asset handle.
+    /// Get an asset and its version from a given asset handle.
     pub fn get_with_version(&self, handle: &Handle<A>) -> Option<&(A, u32)> {
         if self.bitset.contains(handle.id()) {
             Some(unsafe { self.assets.get(handle.id()) })
@@ -145,7 +145,7 @@ impl<A: Asset> AssetStorage<A> {
         }
     }
 
-    /// Get an asset by it's handle id.
+    /// Get an asset by its handle id.
     pub fn get_by_id(&self, id: u32) -> Option<&A> {
         if self.bitset.contains(id) {
             Some(unsafe { &self.assets.get(id).0 })
@@ -193,8 +193,8 @@ impl<A: Asset> AssetStorage<A> {
         self.bitset.contains(id)
     }
 
-    /// Get an asset by it's handle id without checking the internal bitset.
-    /// Use `contains_id` to manually check it's status before access.
+    /// Get an asset by its handle id without checking the internal bitset.
+    /// Use `contains_id` to manually check its status before access.
     ///
     /// # Safety
     /// You must manually verify that given asset id is valid.