+++ amethyst_rendy/src/submodules/gather.rs
@@ -89,7 +89,7 @@ impl CameraGatherer {
 
         let camera_position = convert::<_, Vector3<f32>>(transform.column(3).xyz()).into_pod();
 
-        let proj = camera.as_matrix();
+        let proj = &camera.matrix;
         let view = &**transform;
 
         let proj_view: [[f32; 4]; 4] = ((*proj) * view).into();