+++ examples/debug_lines/main.rs
@@ -1,4 +1,7 @@
-//! Displays several lines with both methods.
+//! Displays dynamic and static debug lines via the resource and component
+//! methods respectively. A pendulum of sorts is drawn using the resource
+//! method, while a grid pattern is drawn on the floor using the component
+//! method.
 
 use amethyst::{
     controls::{FlyControlBundle, FlyControlTag},
@@ -12,7 +15,7 @@ use amethyst::{
     input::{is_close_requested, is_key_down, InputBundle, StringBindings},
     prelude::*,
     renderer::{
-        camera::{Camera, Projection},
+        camera::Camera,
         debug_drawing::{DebugLines, DebugLinesComponent, DebugLinesParams},
         palette::Srgba,
         plugins::{RenderDebugLines, RenderSkybox, RenderToWindow},
@@ -33,18 +36,18 @@ impl<'s> System<'s> for ExampleLinesSystem {
     );
 
     fn run(&mut self, (mut debug_lines_resource, time): Self::SystemData) {
-        // Drawing debug lines, as a resource
+        // Drawing debug lines as a resource
         let t = (time.absolute_time_seconds() as f32).cos();
 
         debug_lines_resource.draw_direction(
-            [t, 0.0, 0.5].into(),
-            [0.0, 0.3, 0.0].into(),
+            Point3::new(t, 0.0, 0.5),
+            Vector3::new(0.0, 0.3, 0.0),
             Srgba::new(0.5, 0.05, 0.65, 1.0),
         );
 
         debug_lines_resource.draw_line(
-            [t, 0.0, 0.5].into(),
-            [0.0, 0.0, 0.2].into(),
+            Point3::new(t, 0.0, 0.5),
+            Point3::new(0.0, 0.0, 0.2),
             Srgba::new(0.5, 0.05, 0.65, 1.0),
         );
     }
@@ -58,21 +61,27 @@ impl SimpleState for ExampleState {
         // Configure width of lines. Optional step
         data.world.insert(DebugLinesParams { line_width: 2.0 });
 
-        // Setup debug lines as a component and add lines to render axis&grid
+        // Setup debug lines as a component and add lines to render axes & grid
         let mut debug_lines_component = DebugLinesComponent::with_capacity(100);
+
+        // X-axis (red)
         debug_lines_component.add_direction(
-            [0.0, 0.0001, 0.0].into(),
-            [0.2, 0.0, 0.0].into(),
+            Point3::new(0.0, 0.0001, 0.0),
+            Vector3::new(0.2, 0.0, 0.0),
             Srgba::new(1.0, 0.0, 0.23, 1.0),
         );
+
+        // Y-axis (yellowish-green)
         debug_lines_component.add_direction(
-            [0.0, 0.0, 0.0].into(),
-            [0.0, 0.2, 0.0].into(),
+            Point3::new(0.0, 0.0, 0.0),
+            Vector3::new(0.0, 0.2, 0.0),
             Srgba::new(0.5, 0.85, 0.1, 1.0),
         );
+
+        // Z-axis (blue)
         debug_lines_component.add_direction(
-            [0.0, 0.0001, 0.0].into(),
-            [0.0, 0.0, 0.2].into(),
+            Point3::new(0.0, 0.0001, 0.0),
+            Vector3::new(0.0, 0.0, 0.2),
             Srgba::new(0.2, 0.75, 0.93, 1.0),
         );
 
@@ -125,6 +134,9 @@ impl SimpleState for ExampleState {
                 }
             }
         }
+
+        // Debug lines are automatically rendered by including the debug lines
+        // rendering plugin
         data.world.register::<DebugLinesComponent>();
         data.world
             .create_entity()
@@ -137,12 +149,11 @@ impl SimpleState for ExampleState {
         data.world
             .create_entity()
             .with(FlyControlTag)
-            .with(Camera::from(Projection::perspective(
+            .with(Camera::perspective(
                 1.33333,
                 std::f32::consts::FRAC_PI_2,
                 0.1,
-                1000.0,
-            )))
+            ))
             .with(local_transform)
             .build();
     }
@@ -171,7 +182,7 @@ fn main() -> amethyst::Result<()> {
 
     let display_config_path = app_root.join("examples/debug_lines/config/display.ron");
     let key_bindings_path = app_root.join("examples/debug_lines/config/input.ron");
-    let assets_dir = app_root.join("examples/assets/");
+    let assets_dir = app_root.join("examples/debug_lines/assets/");
 
     let fly_control_bundle = FlyControlBundle::<StringBindings>::new(
         Some(String::from("move_x")),