+++ examples/pong_tutorial_06/main.rs
@@ -1,0 +1,149 @@
+//! Pong
+
+mod audio;
+mod bundle;
+mod pong;
+mod systems;
+
+use amethyst::{
+    audio::{AudioBundle, DjSystemDesc},
+    core::{frame_limiter::FrameRateLimitStrategy, transform::TransformBundle},
+    ecs::{Component, DenseVecStorage},
+    input::{InputBundle, StringBindings},
+    prelude::*,
+    renderer::{
+        plugins::{RenderFlat2D, RenderToWindow},
+        types::DefaultBackend,
+        RenderingBundle,
+    },
+    ui::{RenderUi, UiBundle},
+    utils::application_root_dir,
+};
+
+use crate::{audio::Music, bundle::PongBundle};
+use std::time::Duration;
+
+const ARENA_HEIGHT: f32 = 100.0;
+const ARENA_WIDTH: f32 = 100.0;
+const PADDLE_HEIGHT: f32 = 16.0;
+const PADDLE_WIDTH: f32 = 4.0;
+const PADDLE_VELOCITY: f32 = 75.0;
+
+const BALL_VELOCITY_X: f32 = 75.0;
+const BALL_VELOCITY_Y: f32 = 50.0;
+const BALL_RADIUS: f32 = 2.0;
+
+const AUDIO_MUSIC: &[&str] = &[
+    "audio/Computer_Music_All-Stars_-_Wheres_My_Jetpack.ogg",
+    "audio/Computer_Music_All-Stars_-_Albatross_v2.ogg",
+];
+const AUDIO_BOUNCE: &str = "audio/bounce.ogg";
+const AUDIO_SCORE: &str = "audio/score.ogg";
+
+fn main() -> amethyst::Result<()> {
+    amethyst::start_logger(Default::default());
+
+    use crate::pong::Pong;
+
+    let app_root = application_root_dir()?;
+
+    let display_config_path = app_root.join("examples/pong_tutorial_06/config/display.ron");
+
+    let key_bindings_path = {
+        if cfg!(feature = "sdl_controller") {
+            app_root.join("examples/pong_tutorial_06/config/input_controller.ron")
+        } else {
+            app_root.join("examples/pong_tutorial_06/config/input.ron")
+        }
+    };
+
+    let assets_dir = app_root.join("examples/pong_tutorial_06/assets/");
+
+    let game_data = GameDataBuilder::default()
+        // Add the transform bundle which handles tracking entity positions
+        .with_bundle(TransformBundle::new())?
+        .with_bundle(
+            InputBundle::<StringBindings>::new().with_bindings_from_file(key_bindings_path)?,
+        )?
+        .with_bundle(PongBundle)?
+        .with_bundle(AudioBundle::default())?
+        .with_system_desc(
+            DjSystemDesc::new(|music: &mut Music| music.music.next()),
+            "dj_system",
+            &[],
+        )
+        .with_bundle(UiBundle::<StringBindings>::new())?
+        .with_bundle(
+            RenderingBundle::<DefaultBackend>::new()
+                // The RenderToWindow plugin provides all the scaffolding for opening a window and
+                // drawing on it
+                .with_plugin(
+                    RenderToWindow::from_config_path(display_config_path)?
+                        .with_clear([0.34, 0.36, 0.52, 1.0]),
+                )
+                .with_plugin(RenderFlat2D::default())
+                .with_plugin(RenderUi::default()),
+        )?;
+
+    let mut game = Application::build(assets_dir, Pong::default())?
+        .with_frame_limit(
+            FrameRateLimitStrategy::SleepAndYield(Duration::from_millis(2)),
+            144,
+        )
+        .build(game_data)?;
+
+    game.run();
+    Ok(())
+}
+
+pub struct Ball {
+    pub velocity: [f32; 2],
+    pub radius: f32,
+}
+
+impl Component for Ball {
+    type Storage = DenseVecStorage<Self>;
+}
+
+#[derive(PartialEq, Eq)]
+pub enum Side {
+    Left,
+    Right,
+}
+
+pub struct Paddle {
+    pub velocity: f32,
+    pub side: Side,
+    pub width: f32,
+    pub height: f32,
+}
+
+impl Paddle {
+    pub fn new(side: Side) -> Paddle {
+        Paddle {
+            velocity: 1.0,
+            side,
+            width: 1.0,
+            height: 1.0,
+        }
+    }
+}
+
+impl Component for Paddle {
+    type Storage = DenseVecStorage<Self>;
+}
+
+#[derive(Default)]
+pub struct ScoreBoard {
+    score_left: i32,
+    score_right: i32,
+}
+
+impl ScoreBoard {
+    pub fn new() -> ScoreBoard {
+        ScoreBoard {
+            score_left: 0,
+            score_right: 0,
+        }
+    }
+}