+++ examples/prefab_adapter/main.rs
@@ -173,7 +173,7 @@ fn main() -> Result<(), Error> {
     let app_root = application_root_dir()?;
 
     // Add our meshes directory to the asset loader.
-    let assets_dir = app_root.join("examples/assets");
+    let assets_dir = app_root.join("examples/prefab_adapter/assets");
 
     let game_data = GameDataBuilder::default().with_system_desc(
         PrefabLoaderSystemDesc::<PositionPrefab>::default(),