+++ docs/CHANGELOG.md
@@ -14,22 +14,38 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 - `GameDataBuilder::build_dispatcher` method returns a standalone `Dispatcher`
   instead of using `DataInit` to build a `GameData` ([#2294])
+- Support for text alignment in `UiButton` and `UiLabel` ([#2316])
+- `amethyst_input::axis::Axis` supports a new variant, `Multiple` ([#2341])
+- Support layer to be set in `UiLabelBuilder` ([#2358])
+- Support line mode to be set in `UiLabelBuilder` and `UiButtonBuilder` ([#2358])
+- `SpriteRender::new` for cleaner instantiation ([#2395])
 
 ### Changed
 
 - `amethyst_rendy::shape::Shape::upload` takes `&ShapeUpload`. ([#2264])
+- Examples now have assets colocated in the individual example directiories ([#2289], [#2305])
+- `UiText` now requires 2 more arguments `line_mode` and `align` ([#2358])
+- `amethyst_ui::UiButtonActionRetrigger` now derives `Default` and `Clone`. ([#2388])
 
 ### Fixed
 
 - Corrected an issue where fixed updates were tied to time scale. ([#2254])
 - Fixed asset handle reuse bug in renderer. ([#2258])
+- Fixed issue where all `TileMap`s were rendered with the same transformation. ([#2210])
 - Fixed UiButtonBuilder incorrect UiImage creation ([#2299])
+- Correctly increment texture generation tracking number. ([#2339])
+- Fixed text alignment never being set in `UiText` ([#2358])
 
+[#2341]: https://github.com/amethyst/amethyst/pull/2341
 [#2294]: https://github.com/amethyst/amethyst/pull/2294
+[#2210]: https://github.com/amethyst/amethyst/issues/2210
 [#2254]: https://github.com/amethyst/amethyst/issues/2254
 [#2258]: https://github.com/amethyst/amethyst/pull/2258
 [#2264]: https://github.com/amethyst/amethyst/pull/2264
 [#2299]: https://github.com/amethyst/amethyst/pull/2299
+[#2339]: https://github.com/amethyst/amethyst/pull/2339
+[#2388]: https://github.com/amethyst/amethyst/pull/2388
+[#2395]: https://github.com/amethyst/amethyst/pull/2395
 
 ## [0.15.0] - 2020-03-24
 
@@ -106,7 +122,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 - amethyst_network completely rewritten to provide a new baseline with which to build. ([#1917])
 - Cleaned up tiles example. Added rotation and translation tests, fixed raycast debug box. Added default zoom to PROJECT
   perspective projection since no one knew to zoom out. ([#1974])
-- TileMaps to_tile and to_world now take an Option<&Transform> that allows them to work if the entire map in 
+- TileMaps to_tile and to_world now take an Option<&Transform> that allows them to work if the entire map in
  translated. ([#1987],[#1991])
 - `AmethystApplication::with_fn` constraint relaxed from `Fn` to `FnOnce`. ([#1983])
 - ScreenDimensions now consistently reports window size in physical pixels. ([#1988])